Refactor motor proto types
