+++ internal/motor/proto/api/v1/request.proto
@@ -1,0 +1,45 @@
+syntax = "proto3";
+
+// Package Motor is used for defining a Motor node and its properties.
+package sonrio.motor.api.v1;
+
+// -----------------------------------------------------------------------------
+// Motor Node API
+// -----------------------------------------------------------------------------
+// (Client) InitializeRequest Message to Establish Sonr Host/API/Room
+message InitializeRequest {
+  string device_id = 1;
+  string home_dir = 2;
+  string support_dir = 3;
+  string temp_dir = 4;
+  HostOptions host_options = 5;         // Libp2p Host config
+
+  // Libp2p Host Options
+  message HostOptions {
+    bool quic_transport = 1;             // Enable QUIC Transport
+    bool http_transport = 2;             // Enable HTTP Transport
+    bool ipv4_only = 3;                  // Enable IPv4 Only
+    repeated IPAddress listen_addrs = 4; // List of Listen Addresses (optional)
+  }
+
+  // IP Address Interface
+  message IPAddress {
+    string name = 1;   // Name of Interface
+    string address = 2;  // IP Address of Interface
+    bool internal = 3; // Wether it is a Loopback Interface
+    Family family = 4; // Address Family
+
+    // Type of IP Interface
+    enum Family {
+      FAMILY_UNSPECIFIED = 0;
+      FAMILY_IPV4 = 1; // IPv4 Address
+      FAMILY_IPV6 = 2; // IPv6 Address
+    }
+  }
+}
+
+// CreateAccount Request contains the three keys needed to create an account on Sonr
+message CreateAccountRequest {
+  string password = 1;
+  bytes aes_dsc_key = 2;
+}
