+++ internal/motor/proto/api/v1/response.proto
@@ -1,0 +1,9 @@
+syntax = "proto3";
+
+// Package Motor is used for defining a Motor node and its properties.
+package sonrio.motor.api.v1;
+
+message CreateAccountResponse {
+  bytes aes_psk = 1;
+  string address = 2;
+}
