+++ internal/motor/proto/buf.yaml
@@ -5,13 +5,16 @@ deps:
   - buf.build/cosmos/cosmos-sdk
 
 lint:
+  allow_comment_ignores: true
   use:
     - DEFAULT
     - COMMENTS
     - FILE_LOWER_SNAKE_CASE
   except:
+    - PACKAGE_DIRECTORY_MATCH
     - UNARY_RPC
     - COMMENT_FIELD
+    - COMMENT_ENUM_VALUE
     - SERVICE_SUFFIX
     - PACKAGE_VERSION_SUFFIX
     - RPC_REQUEST_STANDARD_NAME
