+++ internal/motor/proto/common/v1/info.proto
@@ -1,0 +1,81 @@
+syntax = "proto3";
+
+// Package common defines commonly used types agnostic to the node role on the Sonr network.
+package sonrio.motor.common.v1;
+
+// Direction is the direction of a stream.
+enum Direction {
+  // Unspecified is the default value.
+  DIRECTION_UNSPECIFIED = 0;
+
+  // Incoming is the direction of an incoming stream.
+  DIRECTION_INCOMING = 1;
+
+  // Outgoing is the direction of an outgoing stream.
+  DIRECTION_OUTGOING = 2;
+}
+
+// Standard MIME with Additional Extensions
+message MIME {
+  // Type of File
+  Type type = 1;
+
+  // Extension of File
+  string subtype = 2;
+
+  // Type/Subtype i.e. (image/jpeg)
+  string value = 3;
+
+  // File Content Type
+  enum Type {
+    // Other File Type - If cannot derive from Subtype
+    TYPE_UNSPECIFIED = 0;
+
+    // Sound, Audio Files
+    TYPE_AUDIO = 1;
+
+    // Document Files - PDF, Word, Excel, etc.
+    TYPE_DOCUMENT = 2;
+
+    // Image Files
+    TYPE_IMAGE = 3;
+
+    // Text Based Files
+    TYPE_TEXT = 4;
+
+    // Video Files
+    TYPE_VIDEO = 5;
+
+    // URL Links
+    TYPE_URL = 6;
+
+    // Crypto Files
+    TYPE_CRYPTO = 7;
+  }
+}
+
+// Basic Info Sent to Peers to Establish Connections
+message Peer {
+  string did = 1;       // User Sonr Domain
+  Status status = 2;      // Peer Status
+  Device device = 3;      // Peer Device Info
+  bytes public_key = 4;    // Public Key of the Peer
+  int64 last_modified = 6; // Last Modified Timestamp
+
+  // Peer Info for Device
+  message Device {
+    string id = 1;       // Peer Device ID
+    string host_name = 2; // Peer Host Name
+    string os = 3;       // Peer Operating System
+    string arch = 4;     // Peer Architecture
+    string model = 5;    // Peers Device Model
+  }
+
+  // Peers Active Status
+  enum Status {
+    STATUS_UNSPECIFIED = 0; // Offline - Not Online or Not a Full Node
+    STATUS_ONLINE = 1;  // Online - Full Node Available
+    STATUS_AWAY = 2;    // Away - Not Online, but has a full node
+    STATUS_BUSY = 3;    // Busy - Online, but busy with Transfer
+  }
+}
