--- internal/motor/proto/core/v1/request.proto
@@ -1,90 +1,0 @@
-syntax = "proto3";
-
-// Package Motor is used for defining a Motor node and its properties.
-package sonrio.motor.core.v1;
-import "core/v1/core.proto";
-import "core/v1/data.proto";
-option go_package = "github.com/sonr-io/sonr/pkg/motor";
-
-// -----------------------------------------------------------------------------
-// Motor Node API
-// -----------------------------------------------------------------------------
-// (Client) InitializeRequest Message to Establish Sonr Host/API/Room
-message InitializeRequest {
-  core.v1.Location location = 1;     // Current Runtime Location
-  core.v1.Connection connection = 2; // Internet Connection Type
-  HostOptions host_options = 3;         // Libp2p Host config
-  ServiceOptions service_options = 4;   // Service Config
-  DeviceOptions device_options = 5;     // File System Config
-  map<string,string> variables = 6;     // Domain TXT Records
-
-  // Optional Message to Initialize FileSystem
-  message DeviceOptions {
-    string id = 1;           // Device ID
-    string home_dir = 2;
-    string support_dir = 3;
-    string temp_dir = 4;
-  }
-
-  // Libp2p Host Options
-  message HostOptions {
-    bool quic_transport = 1;             // Enable QUIC Transport
-    bool http_transport = 2;             // Enable HTTP Transport
-    bool ipv4_only = 3;                  // Enable IPv4 Only
-    repeated IPAddress listen_addrs = 4; // List of Listen Addresses (optional)
-  }
-
-  // Service Configuration
-  message ServiceOptions {
-    bool textile = 1;    // Enable Textile Client and Threads
-    bool mailbox = 2;    // Enable Mailbox
-    bool buckets = 3;    // Enable Buckets
-    int32 interval = 4; // Auto Update Interval (seconds) - Default 5s
-  }
-
-  // IP Address Interface
-  message IPAddress {
-    string name = 1;   // Name of Interface
-    string address = 2;  // IP Address of Interface
-    bool internal = 3; // Wether it is a Loopback Interface
-    Family family = 4; // Address Family
-
-    enum Family {
-      FAMILY_UNSPECIFIED = 0;
-      FAMILY_IPV4 = 1; // IPv4 Address
-      FAMILY_IPV6 = 2; // IPv6 Address
-    }
-  }
-}
-
-// (Client) ShareRequest is request to share supplied files/urls with a peer
-message ShareRequest {
-  core.v1.Peer peer = 1; // Peer to Share with
-  repeated core.v1.SupplyItem items = 2; // Supply Items to share
-}
-
-// (Client) DecideRequest is request to respond to a share request
-message DecideRequest {
-  bool decision = 1;       // True if Supply is Active
-  core.v1.Peer peer = 2; // Peer to Share with
-}
-
-// (Client) SearchRequest is Message for Searching for Peer
-message SearchRequest {
-  oneof query {
-    string s_name = 1;  // SName combined with Device ID and Hashed
-    string peer_id = 2; // Peer ID
-  }
-}
-
-message OnLobbyRefreshRequest {}
-
-message OnMailboxMessageRequest {}
-
-message OnTransmitDecisionRequest {}
-
-message OnTransmitInviteRequest {}
-
-message OnTransmitProgressRequest {}
-
-message OnTransmitCompleteRequest {}
