--- internal/motor/proto/exchange/v1/exchange.proto
@@ -1,32 +1,0 @@
-syntax = "proto3";
-
-// Package exchange defines interfaces and types for exchange between two nodes in the network.
-package sonrio.motor.exchange.v1;
-import "core/v1/core.proto";
-import "core/v1/data.proto";
-option go_package = "github.com/sonr-io/sonr/pkg/host/exchange";
-
-message MailboxMessage {
-  string id = 1; // ID is the Message ID
-  core.v1.Payload payload = 2;    // Payload is the message data
-  core.v1.Peer from = 3;         // Users Peer Data
-  core.v1.Peer to = 4;           // Receivers Peer Data
-  core.v1.Metadata metadata = 5; // Metadata
-  int64 created_at = 6; // Timestamp
-}
-
-// Invitation Message sent on RPC
-message InviteRequest {
-  core.v1.Payload payload = 1;   // Attached Data
-  core.v1.Peer from = 3;         // Users Peer Data
-  core.v1.Peer to = 4;           // Receivers Peer Data
-  core.v1.Metadata metadata = 5; // Metadata
-}
-
-// Reply Message sent on RPC
-message InviteResponse {
-  bool decision = 1;               // Success
-  core.v1.Peer from = 3;         // Users Peer Data
-  core.v1.Peer to = 4;           // Receivers Peer Data
-  core.v1.Metadata metadata = 5; // Metadata
-}
