+++ internal/motor/proto/service/v1/discover.proto
@@ -1,0 +1,14 @@
+syntax = "proto3";
+
+// Package Discover is used to find other Peers in the sonr network.
+package sonrio.motor.service.v1;
+import "common/v1/info.proto";
+
+// LobbyMessage is message passed from Peer in Lobby
+message LobbyMessage {
+  common.v1.Peer from = 1;     // Users Peer Data
+  bytes body = 2;    // Payload is the message data
+    bytes signature = 3;             // Signature of the Attached Data
+  int64 created_at = 4; // Timestamp
+  string to = 5; // Target Peer ID, PubSub Topic, etc.
+}
