--- internal/motor/proto/transmit/v1/transmit.proto
@@ -1,34 +1,0 @@
-syntax = "proto3";
-
-// Package Transmit is utilized for direct data transmission between two peers.
-package sonrio.motor.transmit.v1;
-import "core/v1/core.proto";
-import "core/v1/data.proto";
-option go_package = "github.com/sonr-io/sonr/pkg/host/transmit";
-
-message Session {
-  core.v1.Direction direction = 1;
-  core.v1.Peer from = 2;
-  core.v1.Peer to = 3;
-  core.v1.Payload payload = 4;
-  int64 last_updated = 5;
-  repeated SessionItem items = 6;
-  int32 current_index = 7;
-  map<int32, bool> results = 8;
-}
-
-message SessionItem {
-  int32 index = 1;
-  int32 count = 2;
-  core.v1.FileItem item = 3;
-  int64 written = 4;
-  int64 size = 5;
-  int64 total_size = 6;
-  core.v1.Direction direction = 7;
-  string path = 8;
-}
-
-message SessionPayload {
-  core.v1.Payload payload = 1;
-  core.v1.Direction direction = 2;
-}
