+++ pkg/host/host.go
@@ -20,7 +20,6 @@ import (
 	ma "github.com/multiformats/go-multiaddr"
 	// "github.com/pion/webrtc/v3"
 	"github.com/sonr-io/sonr/pkg/config"
-	types "go.buf.build/grpc/go/sonr-io/motor/core/v1"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -32,7 +31,6 @@ type hostImpl struct {
 	events events.EventEmmiter
 
 	// Host and context
-	connection   types.Connection
 	privKey      crypto.PrivKey
 	mdnsPeerChan chan peer.AddrInfo
 	dhtPeerChan  <-chan peer.AddrInfo
